requirements.txt
pytorch.ignite
torch
lasio
pandas
matplotlib
numpy
seaborn